requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
defusedxml==0.7.1
elementpath==4.2.1
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.21
pyOpenSSL==24.0.0
pysaml2==7.5.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
urllib3==2.2.0
Werkzeug==3.0.1
xmlschema==2.5.1
